Refactor modularize
- Modularize graphql query, mutation.
- Modularize models.
- Modularize environment.
- Add helpers module to move problem and cache modules.
