+++ src/graphql/mutation/mod.rs
@@ -1,0 +1,3 @@
+mod account;
+
+pub struct Mutation;
