+++ src/graphql/query/mod.rs
@@ -1,0 +1,2 @@
+mod accounts;
+pub struct Query;
